Refactor EventListContainer: Introduce helper hooks for data management
- Created `useStudyPlanData` for managing study plan fetching and state updates.
- Added `useEnrolledCoursesData` to handle fetching and updating enrolled courses.
- Implemented `useCourseInfoData` for course information fetching and state management.
- Developed `useCourseRatingsData` to manage course ratings data fetching.
- Introduced `useEventListDataManager` as a master hook to coordinate all data fetching and state management.
- Improved code organization, maintainability, and separation of concerns.
- Preserved existing functionality while enhancing documentation and error handling.
